requirements.txt
cachetools==2.0.1
certifi==2018.4.16
chardet==3.0.4
dill==0.2.7.1
Django==2.0.3
django-common-helpers==0.9.1
django-cors-headers==2.2.0
django-cron==0.5.0
django-filter==1.1.0
django-macaddress==1.5.0
django-model-utils==3.1.1
django-storages==1.6.6
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
future==0.16.0
grpcio==1.11.0
httplib2==0.11.3
idna==2.6
netaddr==0.7.19
oauth2client==3.0.0
Pillow==5.1.0
ply==3.8
psutil==5.4.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyJWT==1.6.1
pytz==2018.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
numpy==1.14.2
opencv-python==3.4.0.12
